FIX: Reorganize tests, too
